Initial commit.
Create initial implementation of RFC 6570 URI Template specification.
Add lex module for consuming characters in the input string and a
parse namespace for parsing the literal and expressions within the
template. The parse process returns a parsed template represented
by an ordered collection of literal or expression components. These
are then expanded according to a specified map of variable bindings
in the expansion/expand-template function.