Add parser tests.
Change Parser/can-start? to take a lexer argument instead of a
codepoint.

Add try-parse function which resets the lexer position if the
specified parser fails.

Update the expression parser to return an 'invalid expression'
result if the current expression is empty ({}), unterminated (no
closing }) or cannot be parsed into a varspec list.