Check no varspecs with prefix modifiers are bound to compound values
Update core/template-valid? to check that varspecs with prefix
modifiers are not bound to composite values. This requires the
bindings map to be passed along with the template parse result.

Update the parsed-ok? function to take the template instead of a
section parse result for consistency with the new types-ok?
function.