Split Parser protocol into Parser and Matcher protocols
Add start location to lex buffer which represents the start of
unconsumed input. The current position can be advanced independently
and the range [start, current] represents a segment of the input
that has been matched but not yet consumed by a parser.

Add a Matcher protocol which represents items that advance the
current position but do not consume the input. Change most of the
single-character parsers to be matchers along with the pct-encoded
parser. This gives the low-level items in the grammar the same
type.

The higher-level items in the grammar (varname, modifier) etc. are
still represented by parsers which consume the input buffer. These
use matchers internally to find the extent of their items in the
input.

Add expansion/encode function for encoding all the codepoints within
a lex buffer according to a matcher for items that do not require
encoding. Update the encode-literal and encode-codepoints functions
to use this function.

Document the functions in the lex and parser namespaces.